+++ NEWS.md
@@ -1,6 +1,8 @@
 
 # Development version
 
+* webfakes now compiles on older macOS versions (before 10.12).
+
 # 1.1.1
 
 First release on CRAN